LibWeb: Split GridTrackSizeStyleValue out of StyleValue.{h,cpp}
